Refactor of EnhancedCommandExecutor
