+++ hexstrike_server.py
@@ -178,232 +178,7 @@ cache = HexStrikeCache()
 from core.telemetry_collector import TelemetryCollector
 telemetry = TelemetryCollector()
 
-class EnhancedCommandExecutor:
-    """Enhanced command executor with caching, progress tracking, and better output handling"""
-
-    def __init__(self, command: str, timeout: int = COMMAND_TIMEOUT):
-        self.command = command
-        self.timeout = timeout
-        self.process = None
-        self.stdout_data = ""
-        self.stderr_data = ""
-        self.stdout_thread = None
-        self.stderr_thread = None
-        self.return_code = None
-        self.timed_out = False
-        self.start_time = None
-        self.end_time = None
-
-    def _read_stdout(self):
-        """Thread function to continuously read and display stdout"""
-        if not self.process or not self.process.stdout:
-            return
-        try:
-            for line in iter(self.process.stdout.readline, ''):
-                if line:
-                    self.stdout_data += line
-                    # Real-time output display
-                    logger.info(f"📤 STDOUT: {line.strip()}")
-        except Exception as e:
-            logger.error(f"Error reading stdout: {e}")
-
-    def _read_stderr(self):
-        """Thread function to continuously read and display stderr"""
-        if not self.process or not self.process.stderr:
-            return
-        try:
-            for line in iter(self.process.stderr.readline, ''):
-                if line:
-                    self.stderr_data += line
-                    # Real-time error output display
-                    logger.warning(f"📥 STDERR: {line.strip()}")
-        except Exception as e:
-            logger.error(f"Error reading stderr: {e}")
-
-    def _show_progress(self, duration: float):
-        """Show enhanced progress indication for long-running commands"""
-        if duration > 2:  # Show progress for commands taking more than 2 seconds
-            progress_chars = ModernVisualEngine.PROGRESS_STYLES['dots']
-            start = time.time()
-            i = 0
-            while self.process and self.process.poll() is None:
-                elapsed = time.time() - start
-                char = progress_chars[i % len(progress_chars)]
-
-                # Calculate progress percentage (rough estimate)
-                progress_percent = min((elapsed / self.timeout) * 100, 99.9)
-                progress_fraction = progress_percent / 100
-
-                # Calculate ETA
-                eta = 0
-                if progress_percent > 5:  # Only show ETA after 5% progress
-                    eta = ((elapsed / progress_percent) * 100) - elapsed
-
-                # Calculate speed
-                bytes_processed = len(self.stdout_data) + len(self.stderr_data)
-                speed = f"{bytes_processed/elapsed:.0f} B/s" if elapsed > 0 else "0 B/s"
-
-                # Update process manager with progress
-                ProcessManager.update_process_progress(
-                    self.process.pid,
-                    progress_fraction,
-                    f"Running for {elapsed:.1f}s",
-                    bytes_processed
-                )
-
-                # Create beautiful progress bar using ModernVisualEngine
-                progress_bar = ModernVisualEngine.render_progress_bar(
-                    progress_fraction,
-                    width=30,
-                    style='cyber',
-                    label=f"⚡ PROGRESS {char}",
-                    eta=eta,
-                    speed=speed
-                )
-
-                logger.info(f"{progress_bar} | {elapsed:.1f}s | PID: {self.process.pid}")
-                time.sleep(0.8)
-                i += 1
-                if elapsed > self.timeout:
-                    break
-
-    def execute(self) -> Dict[str, Any]:
-        """Execute the command with enhanced monitoring and output"""
-        self.start_time = time.time()
-
-        logger.info(f"🚀 EXECUTING: {self.command}")
-        logger.info(f"⏱️  TIMEOUT: {self.timeout}s | PID: Starting...")
-
-        try:
-            self.process = subprocess.Popen(
-                self.command,
-                shell=True,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                bufsize=1
-            )
-
-            pid = self.process.pid
-            logger.info(f"🆔 PROCESS: PID {pid} started")
-
-            # Register process with ProcessManager (v5.0 enhancement)
-            ProcessManager.register_process(pid, self.command, self.process)
-
-            # Start threads to read output continuously
-            self.stdout_thread = threading.Thread(target=self._read_stdout)
-            self.stderr_thread = threading.Thread(target=self._read_stderr)
-            self.stdout_thread.daemon = True
-            self.stderr_thread.daemon = True
-            self.stdout_thread.start()
-            self.stderr_thread.start()
-
-            # Start progress tracking in a separate thread
-            progress_thread = threading.Thread(target=self._show_progress, args=(self.timeout,))
-            progress_thread.daemon = True
-            progress_thread.start()
-
-            # Wait for the process to complete or timeout
-            try:
-                self.return_code = self.process.wait(timeout=self.timeout)
-                self.end_time = time.time()
-
-                # Process completed, join the threads
-                self.stdout_thread.join(timeout=1)
-                self.stderr_thread.join(timeout=1)
-
-                execution_time = self.end_time - self.start_time
-
-                # Cleanup process from registry (v5.0 enhancement)
-                ProcessManager.cleanup_process(pid)
-
-                if self.return_code == 0:
-                    logger.info(f"✅ SUCCESS: Command completed | Exit Code: {self.return_code} | Duration: {execution_time:.2f}s")
-                    telemetry.record_execution(True, execution_time)
-                else:
-                    logger.warning(f"⚠️  WARNING: Command completed with errors | Exit Code: {self.return_code} | Duration: {execution_time:.2f}s")
-                    telemetry.record_execution(False, execution_time)
-
-            except subprocess.TimeoutExpired:
-                self.end_time = time.time()
-                execution_time = self.end_time - self.start_time
-
-                # Process timed out but we might have partial results
-                self.timed_out = True
-                logger.warning(f"⏰ TIMEOUT: Command timed out after {self.timeout}s | Terminating PID {self.process.pid}")
-
-                # Try to terminate gracefully first
-                self.process.terminate()
-                try:
-                    self.process.wait(timeout=5)
-                except subprocess.TimeoutExpired:
-                    # Force kill if it doesn't terminate
-                    logger.error(f"🔪 FORCE KILL: Process {self.process.pid} not responding to termination")
-                    self.process.kill()
-
-                self.return_code = -1
-                telemetry.record_execution(False, execution_time)
-
-            # Always consider it a success if we have output, even with timeout
-            success = True if self.timed_out and (self.stdout_data or self.stderr_data) else (self.return_code == 0)
-
-            # Log enhanced final results with summary using ModernVisualEngine
-            output_size = len(self.stdout_data) + len(self.stderr_data)
-            execution_time = self.end_time - self.start_time if self.end_time else 0
-
-            # Create status summary
-            status_icon = "✅" if success else "❌"
-            status_color = ModernVisualEngine.COLORS['MATRIX_GREEN'] if success else ModernVisualEngine.COLORS['HACKER_RED']
-            timeout_status = f" {ModernVisualEngine.COLORS['WARNING']}[TIMEOUT]{ModernVisualEngine.COLORS['RESET']}" if self.timed_out else ""
-
-            # Create beautiful results summary
-            results_summary = f"""
-{ModernVisualEngine.COLORS['MATRIX_GREEN']}{ModernVisualEngine.COLORS['BOLD']}╭─────────────────────────────────────────────────────────────────────────────╮{ModernVisualEngine.COLORS['RESET']}
-{ModernVisualEngine.COLORS['BOLD']}│{ModernVisualEngine.COLORS['RESET']} {status_color}📊 FINAL RESULTS {status_icon}{ModernVisualEngine.COLORS['RESET']}
-{ModernVisualEngine.COLORS['BOLD']}├─────────────────────────────────────────────────────────────────────────────┤{ModernVisualEngine.COLORS['RESET']}
-{ModernVisualEngine.COLORS['BOLD']}│{ModernVisualEngine.COLORS['RESET']} {ModernVisualEngine.COLORS['NEON_BLUE']}🚀 Command:{ModernVisualEngine.COLORS['RESET']} {self.command[:55]}{'...' if len(self.command) > 55 else ''}
-{ModernVisualEngine.COLORS['BOLD']}│{ModernVisualEngine.COLORS['RESET']} {ModernVisualEngine.COLORS['CYBER_ORANGE']}⏱️  Duration:{ModernVisualEngine.COLORS['RESET']} {execution_time:.2f}s{timeout_status}
-{ModernVisualEngine.COLORS['BOLD']}│{ModernVisualEngine.COLORS['RESET']} {ModernVisualEngine.COLORS['WARNING']}📊 Output Size:{ModernVisualEngine.COLORS['RESET']} {output_size} bytes
-{ModernVisualEngine.COLORS['BOLD']}│{ModernVisualEngine.COLORS['RESET']} {ModernVisualEngine.COLORS['ELECTRIC_PURPLE']}🔢 Exit Code:{ModernVisualEngine.COLORS['RESET']} {self.return_code}
-{ModernVisualEngine.COLORS['BOLD']}│{ModernVisualEngine.COLORS['RESET']} {status_color}📈 Status:{ModernVisualEngine.COLORS['RESET']} {'SUCCESS' if success else 'FAILED'} | Cached: Yes
-{ModernVisualEngine.COLORS['MATRIX_GREEN']}{ModernVisualEngine.COLORS['BOLD']}╰─────────────────────────────────────────────────────────────────────────────╯{ModernVisualEngine.COLORS['RESET']}
-"""
-
-            # Log the beautiful summary
-            for line in results_summary.strip().split('\n'):
-                if line.strip():
-                    logger.info(line)
-
-            return {
-                "stdout": self.stdout_data,
-                "stderr": self.stderr_data,
-                "return_code": self.return_code,
-                "success": success,
-                "timed_out": self.timed_out,
-                "partial_results": self.timed_out and (self.stdout_data or self.stderr_data),
-                "execution_time": self.end_time - self.start_time if self.end_time else 0,
-                "timestamp": datetime.now().isoformat()
-            }
-
-        except Exception as e:
-            self.end_time = time.time()
-            execution_time = self.end_time - self.start_time if self.start_time else 0
-
-            logger.error(f"💥 ERROR: Command execution failed: {str(e)}")
-            logger.error(f"🔍 TRACEBACK: {traceback.format_exc()}")
-            telemetry.record_execution(False, execution_time)
-
-            return {
-                "stdout": self.stdout_data,
-                "stderr": f"Error executing command: {str(e)}\n{self.stderr_data}",
-                "return_code": -1,
-                "success": False,
-                "timed_out": False,
-                "partial_results": bool(self.stdout_data or self.stderr_data),
-                "execution_time": execution_time,
-                "timestamp": datetime.now().isoformat()
-            }
-
+from core.enhanced_command_executor import EnhancedCommandExecutor
 from core.ai_exploit_generator import AIExploitGenerator
 
 class VulnerabilityCorrelator:
